pyproject.toml
[project]
name = "example-drf-movies-catalog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "markdown>=3.8",
    "psycopg>=3.2.6",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "psycopg[binary]>=3.2.6",
    "ruff>=0.11.5",
]

# ruff settings

[tool.ruff]
required-version = ">=0.11.5"

src = [
    "movies_catalog",
]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    ".venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.13
target-version = "py313"

[tool.ruff.lint]
select = [
    "FAST", # FastAPI (FAST)
    "YTT", # flake8-2020 (YTT)
    "ANN", # flake8-annotations (ANN)
    # "ASYNC", # flake8-async (ASYNC)
    "S", # flake8-bandit (S)
    "BLE", # flake8-blind-except (BLE)
    "FBT", # flake8-boolean-trap (FBT)
    "B", # flake8-bugbear (B)
    "COM", # flake8-commas (COM)
    "C4", # flake8-comprehensions (C4)
    "DTZ", # flake8-datetimez (DTZ)
    "T10", # flake8-debugger (T10)
    "DJ", # flake8-django (DJ)
    "EM", # flake8-errmsg (EM)
    "ISC", # flake8-implicit-str-concat (ISC)
    "ICN", # flake8-import-conventions (ICN)
    "LOG", # flake8-logging (LOG)
    "G", # flake8-logging-format (G)
    "INP", # flake8-no-pep420 (INP)
    "PIE", # flake8-pie (PIE)
    "T20", # flake8-print (T20)
    "PYI", # flake8-pyi (PYI)
    # "PT", # flake8-pytest-style (PT)
    "Q", # flake8-quotes (Q)
    "RSE", # flake8-raise (RSE)
    "RET", # flake8-return (RET)
    "SLF", # flake8-self (SLF)
    "SIM", # flake8-simplify (SIM)
    "SLOT", # flake8-slots (SLOT)
    "TC", # flake8-type-checking (TC)
    "ARG", # flake8-unused-arguments (ARG)
    "PTH", # flake8-use-pathlib (PTH)
    "FLY", # flynt (FLY)
    "I", # isort (I)
    "C90", # mccabe (C90)
    "N", # pep8-naming (N)
    "PERF", # Perflint (PERF)
    "E", # pycodestyle Error (E)
    "W", # pycodestyle Warning (W)
    # "DOC", # pydoclint (DOC)
    # "D", # pydocstyle (D)
    "F", # Pyflakes (F)
    "PGH", # pygrep-hooks (PGH)
    "PLC", # Pylint Convention (PLC)
    "PLE", # Pylint Error (PLE)
    "PLR", # Pylint Refactor (PLR)
    "PLW", # Pylint Warning (PLW)
    "UP", # pyupgrade (UP)
    "FURB", # refurb (FURB)
    "RUF", # Ruff-specific rules (RUF)
    "TRY", # tryceratops (TRY)
]
ignore = [
    "S101", # S101	assert	Use of assert detected
    "RUF001", # ambiguous-unicode-character-string (RUF001)
    "RUF002", # ambiguous-unicode-character-docstring (RUF002)
    "RUF003", # ambiguous-unicode-character-comment (RUF003)
    "TC006", # runtime-cast-value (TC006)
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.extend-per-file-ignores]
"**/models/__init__.py" = [
    "PLC0414", # PLC0414 Import alias does not rename original package
]
"**/migrations/*.py" = [
    "RUF012", # RUF012 Mutable class attributes should be annotated with `typing.ClassVar`
]
"movies_catalog/movies_catalog/settings.py" = [
    "S105", # S105 Possible hardcoded password assigned to: "SECRET_KEY"
    "PLW1508", # PLW1508 Invalid type for environment variable default; expected `str` or `None`
    "E501", # E501 Line too long (91 > 88)
]
